Create useGetClientId hook
